realtek-poe: De-suckify "poe sendframe" command
Typing "0x" repeatedly in a string of numbers that's obviously
hexadecimal is painful. So don't use sscanf(), which forces the "0x"
prefix. Instead, use strtoul() in base 16. This accepts numbers with
and without the prefix.

Also, if the string we receive is crap for whatever reason, don't try
to send it to the PoE controller. Just throw it the fuck out.