realtek-poe: Use read16_be()/write16_be() serialization helpers
Why write an integer byte by byte and manually scale it? For just
free.99, you get a helper to read it for you. I'll include the write
helper for free, just pay separate handling and shipping.

Although there is no net reduction in lines of code, the functions
dealing with multi-byte integers are easier to comperstand.