realtek-poe: Improve error messages for invalid port settings
When realtek-poe finds a port with settings it doesn't like, it prints
a cryptic message and continues on. The only indication that something
is wrong might be a non-working port, or no telemetry for a port.

	"invalid port settings"
	"invalid port id"

Are very poor messages. The only way to tell what went wrong is to
turn on hardcore hacking mode. This is not user-friendly.

Change these mssages to something more descriptive. If the port name
is available, include it in the error message.