realtek-poe: Explicitly disable MCU port mapping
Currently, realtek-poe works by interpreting the port ID to map to a
specific port name. This works when there is a 1:1 mapping between the
port ID and PSE ID.

The MCU has the ability to alter the port ID to PSE mapping. In that
case the port IDs that realtek-poe interprets can be wrong. We can't
guarantee that the MCU will not apply such a map, and thus it is
prudent to explicitly disable it.

Note that the mapping seems to have no effect on power delivery. It
only changes the port IDs used in the serial protocol. Explicitly
disable this mapping by sending the "Set port map disable" command.